Implementing database and di
Saves results from previous sessions using a database to store all relevant data. AppDbContext.cs serves as a bridge between our game app and the database where the information is collected. New updates to GameResultService.cs that now has four functions that are responsible for managing data elements in the database. It works but some changes still have to be made for a complete functionality.